requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
langchain-text-splitters
chromadb
pdfplumber
requests
python-multipart
pypdf
python-docx
pypdf
python-docx
